+++ src/listener.c
@@ -45,7 +45,7 @@ static struct bind_kw_list bind_keywords = {
 /* list of the temporarily limited listeners because of lack of resource */
 static struct mt_list global_listener_queue = MT_LIST_HEAD_INIT(global_listener_queue);
 static struct task *global_listener_queue_task;
-static struct task *manage_global_listener_queue(struct task *t, void *context, unsigned short state);
+static struct task *manage_global_listener_queue(struct task *t, void *context, unsigned int state);
 
 
 #if defined(USE_THREAD)
@@ -113,7 +113,7 @@ int accept_queue_push_mp(struct accept_queue_ring *ring, struct connection *conn
 }
 
 /* proceed with accepting new connections */
-static struct task *accept_queue_process(struct task *t, void *context, unsigned short state)
+static struct task *accept_queue_process(struct task *t, void *context, unsigned int state)
 {
 	struct accept_queue_ring *ring = context;
 	struct connection *conn;
@@ -1148,7 +1148,7 @@ REGISTER_POST_DEINIT(listener_queue_deinit);
  * for global resources when there are enough free resource, or at least once in
  * a while. It is designed to be called as a task.
  */
-static struct task *manage_global_listener_queue(struct task *t, void *context, unsigned short state)
+static struct task *manage_global_listener_queue(struct task *t, void *context, unsigned int state)
 {
 	/* If there are still too many concurrent connections, let's wait for
 	 * some of them to go away. We don't need to re-arm the timer because
